+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 3.0.3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix long-standing bug of setting an ``int`` as return value
+  for the transaction manager's ``sortKey`` method. It must be a string.
 
 
 3.0.2 (2018-03-16)